move tests from locobot folder into droidlet folder
[ghstack-poisoned]
