+++ .gitignore
@@ -42,3 +42,8 @@ chat.lock
 incoming_chat.txt
 locobot/test/test_assets/perception_handlers/
 locobot/agent/dashboard_data.db*
+nsp_outputs.csv
+agents/locobot/models
+agents/locobot/datasets
+agents/locobot/dashboard_data.db
+droidlet/perception/tests/test_assets/perception_handlers
