+++ droidlet/interpreter/tests/base_fakeagent_test_case.py
@@ -6,8 +6,8 @@ from unittest.mock import Mock
 import copy
 
 from base_agent.dialogue_objects import AwaitResponse
-from locobot.test.fake_agent import FakeAgent
-from locobot.test.world import World, Opt, SimpleHuman, make_human_opts
+from .fake_agent import FakeAgent
+from .world import World, Opt, SimpleHuman, make_human_opts
 
 
 class BaseFakeAgentTestCase(unittest.TestCase):
