+++ droidlet/interpreter/tests/test_interpreter_mock.py
@@ -5,10 +5,10 @@ import unittest
 
 import numpy as np
 
-from locobot.test.base_fakeagent_test_case import BaseFakeAgentTestCase
-import locobot.agent.rotation as rotation
+from .base_fakeagent_test_case import BaseFakeAgentTestCase
+import droidlet.lowlevel.locobot.rotation as rotation
 from base_agent.test.all_test_commands import MOVE_COMMANDS, GROUND_TRUTH_PARSES
-from locobot.test.test_utils import assert_turn_degree
+from droidlet.lowlevel.locobot.tests.test_utils import assert_turn_degree
 
 CUBE1 = (9, 0, 4)
 CUBE2 = (9, 0, 10)
