+++ droidlet/lowlevel/locobot/tests/smoke_test.py
