+++ droidlet/lowlevel/locobot/tests/test_assets/default.phys_scene_config.json
