+++ droidlet/lowlevel/locobot/tests/test_assets/expect/go_to_absolute1.png
