+++ droidlet/lowlevel/locobot/tests/test_assets/expect/go_to_absolute2.png
