+++ droidlet/lowlevel/locobot/tests/test_assets/human_female.object_config.json
