+++ droidlet/lowlevel/locobot/tests/test_assets/human_female.phys_properties.json
