+++ droidlet/lowlevel/locobot/tests/test_assets/human_male.object_config.json
