+++ droidlet/lowlevel/locobot/tests/test_assets/human_male.phys_properties.json
