+++ droidlet/lowlevel/locobot/tests/test_assets/rp_mei_posed_scaled.glb
