+++ droidlet/lowlevel/locobot/tests/test_mover.py
@@ -4,8 +4,8 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import unittest
 import os
 import sys
-from locobot.test.test_utils import assert_distance_moved, assert_turn_degree
-from locobot.agent.locobot_mover import LoCoBotMover
+from .test_utils import assert_distance_moved, assert_turn_degree
+from droidlet.lowlevel.locobot.locobot_mover import LoCoBotMover
 
 IP = "127.0.0.1"
 if os.getenv("LOCOBOT_IP"):
