--- locobot/incoming_chat.txt
@@ -1,1 +1,0 @@
-1 | <commandline_speaker> hi
