+++ tools/data_scripts/fetch_aws_models.sh
@@ -49,5 +49,5 @@ tar -xzvf $MODELS_DIRNAME.tar.gz -C $AGENT/agent/models/ --strip-components 1 ||
 if [ $AGENT == "locobot" ]; then
     curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_models.tar.gz -o locobot_models.tar.gz
     tar -xzvf locobot_models.tar.gz -C $AGENT/agent/models/
-	curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_test_assets.tar.gz | tar -xzv -C $AGENT/test/test_assets/
+	curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_test_assets.tar.gz | tar -xzv -C droidlet/perception/tests/test_assets/
 fi
